refactor: move SpinalSimApbI2C
